+++ src/main/java/co/nvqa/common/core/cucumber/glue/ApiOrderSteps.java
@@ -5,6 +5,7 @@ import co.nvqa.common.core.client.OrderClient;
 import co.nvqa.common.core.cucumber.CoreStandardSteps;
 import co.nvqa.common.core.model.EditDeliveryOrderRequest;
 import co.nvqa.common.core.model.Lazada3PL;
+import co.nvqa.common.core.model.order.BulkForceSuccessOrderRequest;
 import co.nvqa.common.core.model.order.Order;
 import co.nvqa.common.core.model.order.Order.Dimension;
 import co.nvqa.common.core.model.order.RescheduleOrderRequest;
@@ -189,6 +190,25 @@ public class ApiOrderSteps extends CoreStandardSteps {
     final boolean codCollected = Boolean.parseBoolean(codCollectedString);
     doWithRetry(
         () -> getOrderClient().forceSuccess(orderId, codCollected), "Force success", 3000, 10);
+  }
+
+  /**
+   * API Core - Operator bulk force success all orders with cod collected "false"
+   *
+   * @param codCollected example: false
+   */
+  @When("API Core - Operator bulk force success below orders with cod collected {string}:")
+  public void apiOperatorBulkForceSuccessAllOrders(String codCollected, List<String> orderIds) {
+    final List<String> finalOrderIds = resolveValues(orderIds);
+    List<BulkForceSuccessOrderRequest> request = new ArrayList<>();
+    finalOrderIds.stream().distinct().forEach(e -> {
+      BulkForceSuccessOrderRequest forceSuccessRequest = new BulkForceSuccessOrderRequest();
+      forceSuccessRequest.setOrderId(Long.parseLong(e));
+      forceSuccessRequest.setIsCodCollected(Boolean.valueOf(codCollected));
+      forceSuccessRequest.setReason("QA AUTO TEST BULK FORCE SUCCESS");
+      request.add(forceSuccessRequest);
+    });
+    getOrderClient().bulkForceSuccess(request);
   }
 
   /**