+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Change log
+<hr>
+
+## [1.3.17] 2023-08-08
+
+### Added
+- RouteResponse.java - Add Hub to represent the hub associated with the route
+- API Core - Operator bulk force success below orders with cod collected {string}:
+- API Core - Operator force success waypoint with cod collected as {string} using route manifest:
+- API Route - Operator run FM PAJ auto route cron job for date {string}
 
 <hr>